tests/test_dynamic_interaction_plugins.py
import unittest
from selenium.webdriver import Chrome, ChromeOptions
from extensions.dynamic_interaction_plugins import DynamicInteractionPlugin
from scraper.dynamic_content_handler import DynamicContentHandler
from utilities.logger import Logger

class TestDynamicInteractionPlugins(unittest.TestCase):

    def setUp(self):
        # Set up the Chrome WebDriver and Logger for testing
        options = ChromeOptions()
        options.headless = True  # Run in headless mode for tests
        self.driver = Chrome(options=options)
        self.logger = Logger("TestDynamicInteractionPlugins")
        self.content_handler = DynamicContentHandler(self.driver, self.logger)
        self.plugin = DynamicInteractionPlugin(self.content_handler)

    def tearDown(self):
        # Close the WebDriver after each test
        self.driver.quit()

    def test_plugin_initialization(self):
        # Test the initialization of the DynamicInteractionPlugin
        self.assertIsInstance(self.plugin, DynamicInteractionPlugin, "The plugin should be an instance of DynamicInteractionPlugin")

    def test_plugin_interaction(self):
        # Test a specific interaction provided by the plugin
        # This is a placeholder test and should be replaced with actual interaction tests
        try:
            self.plugin.perform_interaction()
            self.assertTrue(True, "The interaction should be performed without exceptions")
        except Exception as e:
            self.fail(f"Interaction raised an exception: {e}")

    # Additional tests for other interactions and edge cases should be added here

if __name__ == '__main__':
    unittest.main()

